+++ .github/workflows/phar.yml
@@ -18,7 +18,7 @@ jobs:
 
     strategy:
       matrix:
-        php: [5.6]
+        php: [7.0]
         setup: [prefer-stable]
 
     name: Release phar
